// src/app/interfaces/userDataToken.interface.ts
export interface UserDataToken {
  id: string
  uid: string
  name: string
  secondName: string
  email: string
  lastname: string
  secondLastname: string
  token: string
  roles: string[]
  phoneNumber: string
  birthdate: string
  documentNumber: string
  documentType: string
  commission: string
  identificationNumber: string
}